reorg nodes
